requirements.txt
numpy
pandas
torch
tranformers
sentence-transformers
fastapi
kaleido
python-multipart
-e git+https://github.com/hypercycle-development/pyhypercycle-aim.git#egg=pyhypercycle_aim